refactor: modal 부분 파일명 변경 및 폴더 구조 수정
