+++ src/components/molecules/CategoryMoribSet/index.tsx
@@ -1,9 +1,9 @@
 import { useRef } from 'react';
 
+import AddCategoryListModal from '@/components/AddCategoryListModal';
 import CategoryInputTitle from '@/components/atoms/CategoryInputTitle/index';
 import CategoryUrlInput from '@/components/atoms/CategoryUrlInput/index';
 import GetCategoryBtn from '@/components/atoms/GetCategoryBtn/index';
-import AddCategoryListModal from '@/components/templates/AddCategoryListModal';
 
 interface UrlInfo {
 	url: string;
